feat: classes 클래스 생성 모달 에러 해결
