+++ src/components/class/ClassFrame.jsx
@@ -1,7 +1,3 @@
-import { useParams } from "react-router-dom";
-import { useDispatch } from "react-redux";
-import { useEffect } from "react";
-import { __getClassInfo } from "../../modules/redux/cookingClass";
 import styled from "styled-components";
 
 import GridTemplate from "../../elements/templates/GridTemplate";
@@ -9,23 +5,15 @@ import ChatFrame from "./ChatFrame";
 // import VideoFrame from "./VideoFrame";
 
 const ClassFrame = () => {
-  const { classId, redisClassId, role } = useParams();
-
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(__getClassInfo({ classId }));
-  }, [dispatch, classId]);
-
   return (
     <GridTemplate>
       <StClassSection>
         <div>classSection</div>
         {/* <VideoFrame /> */}
       </StClassSection>
-      {/* <StChatSection>
+      <StChatSection>
         <ChatFrame />
-      </StChatSection> */}
+      </StChatSection>
     </GridTemplate>
   );
 };
