+++ src/components/classes/modal/CreateModal.jsx
@@ -28,12 +28,12 @@ const CreateModal = ({ onClick }) => {
 
   const {
     register,
+    setValue,
     handleSubmit,
     formState: { errors },
     resetField,
   } = useForm({
     mode: "onChange",
-    defaultValues: { recipe: selectedRecipe.name },
   });
 
   // 사용자가 모달창 각 인풋필드 클릭 시 동작
@@ -80,7 +80,7 @@ const CreateModal = ({ onClick }) => {
   };
 
   // 클래스 생성
-  const submitHandler = async ({ className, classImgs, _ }) => {
+  const submitHandler = async ({ className, classImgs, recipe }) => {
     const resp = await apis.create_class({
       recipe_id: selectedRecipe.id,
       class_name: className,
@@ -92,7 +92,6 @@ const CreateModal = ({ onClick }) => {
       status: { code },
     } = resp.data;
 
-    // TODO 썸네일 크기 메시지 추가
     if (code === 400) {
       errors.classImg = {
         message: "썸네일 크기는 20MB를 넘을 수 없습니다.",
@@ -138,10 +137,9 @@ const CreateModal = ({ onClick }) => {
                 id='recipe'
                 onClick={clickHandler}
                 onChange={clickHandler}
-                value={selectedRecipe?.recipe_name || ""}
+                value={selectedRecipe.recipe_name || ""}
                 {...register("recipe", {
-                  required:
-                    "신규 클래스 생성을 위해서 해당하는 레시피를 입력해주셔야해요.",
+                  required: "신규 클래스 생성을 위해서 레시피를 입력해주세요.",
                 })}
               />
               {errors.recipe ? <ET1>{errors.recipe.message}</ET1> : null}
