+++ src/modules/redux/cookingClass.jsx
@@ -10,29 +10,23 @@ const initialState = {
   viewerNum: 0,
   isLoading: false,
   error: "",
-  stomp: {
-    stompClient: {},
-    // getData: () => {},
-    // creatData: () => {},
-    // sendData: () => {},
-  },
 };
 
-export const __getClassInfo = createAsyncThunk(
-  "cookingClass/__getClassInfo",
-  async ({ classId }, thunkAPI) => {
-    try {
-      const resp = await apis.get_class_info({ classId });
-      const {
-        content: { session_id, token, chats },
-      } = resp.data;
+// export const __getClassInfo = createAsyncThunk(
+//   "cookingClass/__getClassInfo",
+//   async ({ classId }, thunkAPI) => {
+//     try {
+//       const resp = await apis.get_class_info({ classId });
+//       const {
+//         content: { session_id, token, chats },
+//       } = resp.data;
 
-      return thunkAPI.fulfillWithValue({ session_id, token, chats });
-    } catch (e) {
-      return thunkAPI.rejectWithValue(e.code);
-    }
-  }
-);
+//       return thunkAPI.fulfillWithValue({ session_id, token, chats });
+//     } catch (e) {
+//       return thunkAPI.rejectWithValue(e.code);
+//     }
+//   }
+// );
 
 const cookingClassSlice = createSlice({
   name: "cookingClass",
@@ -61,40 +55,25 @@ const cookingClassSlice = createSlice({
       const { message } = action.payload;
       state.prevChats = [...state.prevChats, message];
     },
-    saveStompClient: (state, action) => {
-      const { stompClient, getHeader, createData, sendEvent } = action.payload;
-      state.stomp = {
-        stompClient,
-        getHeader,
-        createData,
-        sendEvent,
-      };
-    },
   },
   extraReducers: {
-    [__getClassInfo.pending]: (state, _) => {
-      state.isLoading = true;
-    },
-    [__getClassInfo.fulfilled]: (state, action) => {
-      state.isLoading = false;
-      const { session_id, token, chats } = action.payload;
-      state.prevChats = chats;
-      state.sessionId = session_id;
-      state.token = token;
-    },
-    [__getClassInfo.rejected]: (state, action) => {
-      state.isLoading = false;
-      state.error = action.payload;
-    },
+    // [__getClassInfo.pending]: (state, _) => {
+    //   state.isLoading = true;
+    // },
+    // [__getClassInfo.fulfilled]: (state, action) => {
+    //   state.isLoading = false;
+    //   const { session_id, token, chats } = action.payload;
+    //   state.prevChats = chats;
+    //   state.sessionId = session_id;
+    //   state.token = token;
+    // },
+    // [__getClassInfo.rejected]: (state, action) => {
+    //   state.isLoading = false;
+    //   state.error = action.payload;
+    // },
   },
 });
 
-export const {
-  openModal,
-  closeModal,
-  enterClass,
-  createdClass,
-  sendMessage,
-  saveStompClient,
-} = cookingClassSlice.actions;
+export const { openModal, closeModal, enterClass, createdClass, sendMessage } =
+  cookingClassSlice.actions;
 export default cookingClassSlice.reducer;
